Move Language enum into language-core
Was in market-core before, but turns out it's market that needs to know
about Language, Language itself and LanguageService do not need to know
anything about Market.
This enables us to while setting a new market, to also make sure that we
set the right language at the same time.
