+++ app/feature/feature-profile/src/main/kotlin/com/hedvig/android/feature/profile/settings/SettingsViewModel.kt
@@ -2,8 +2,8 @@ package com.hedvig.android.feature.profile.settings
 
 import com.hedvig.android.data.settings.datastore.SettingsDataStore
 import com.hedvig.android.hanalytics.featureflags.FeatureManager
+import com.hedvig.android.language.Language
 import com.hedvig.android.language.LanguageService
-import com.hedvig.android.market.Language
 import com.hedvig.android.market.Market
 import com.hedvig.android.market.MarketManager
 import com.hedvig.android.memberreminders.EnableNotificationsReminderManager
